Initial image cutout service framework
This has all of the machinery for an image cutout service without
the image cutout backend, only a worker that returns dummy data.
It does implement cutout parameter parsing, the full IVOA UWS job
system, and the required SODA endpoints including VOSI.

The intent is for the contents of the uws directory to move into
Safir (and their dependencies as an optional safir dependency, such
as safir[uws]), so both those modules and their tests should not
require anything else in the image cutout application except for
auth_dependency, which should also move into Safir.

See https://dmtn-208.lsst.io/v/DM-31792/index.html for discussion
of the implementation strategy.