Rework event model, add Butler, add small backend
Incorporate the draft cutout backend from @SimonKrughoff, although
assume a simplistic conversion from degrees to pixels that would
not be acceptable for the production service.  This is intended to
be enough for end-to-end testing.

Redo the event scheme so that UWS database updates are handled via
different actors than the cutout itself.  These actors can share
source code with the frontend and are called either explicitly or
using the Callbacks middleware.  This enables isolating the code that
requires the stack from all other infrastructure.

Isolate the cutout actor in its own file, which also redeclares the
broker and middleware and stubs for other actors.  This file will
be mounted into a stack container and used to start the cutout
backend.

Rework the error representation to explicitly include error codes in
the metadata, since this is a closer match to the error message model
in the SODA standard.  Rework the UWS exception hierarchy accordingly.

Instead of returning a URL from the cutout actor, return the data
required to request the URL from Butler.  Update the database schema
accordingly.  Retrieve the URL from Butler on the fly in the frontend,
which will be required in the future to support signed URLs (since
those URLs will expire).