Add script to start image
Use a script to start the image instead of running uvicorn directly,
and use that to install the database schema first.  Add a click
CLI for vo-cutouts and use that to implement the init command to
install the database schema.  Fix the entry point definition to
define a vo-cutouts command, not vocutouts.