Convert visit and detector to integers
Butler appears to need these as integers and not strings.
Don't put quotes around the numeric values.  This doesn't work in
SQL and presumably won't work for Butler.

Add some delays to job failure tests so that we have some hope of
the started message being processed before the failed message.