Make tests robust against out of order messages
Despite message prioritization there is still a race condition in
the tests where the starting message may be processed after the
job completion message, which means the job status changes before
the start time field is populated.  When waiting for jobs to finish
in the test suite, wait until the start time field is populated as
well.