Overhaul backend integration
Replace the temporary initial backend that uses subprocess to run
pipetask with a proper backend based on lsst.image_cutout_backend.

The new backend returns an S3 URL where the result is stored rather
than using an output Butler collection, which removes the need for
Butler support in the frontend.  (We were going to use it to check
the validity of the data ID, but do not do this for now until Butler
becomes easier to integrate with a Python project.)  Remove Butler
from the UWS code and configuration, and replace it with a simpler
layer around Google Cloud Storage.

The new backend does not (yet) support POS RANGE, so replace the
tests with POS CIRCLE instead.

Log exceptions when dispatching a request to Dramatiq, since for
some reason an exception thrown there is never logged and only
produces an uninformative 500 error.

Thanks to eliminating Butler, we can return to pinning dependencies
since we no longer have to depend on the Git version of Butler.

The UWS database schema has changed since job results no longer
store a set of Butler parameters, and instead only have to store a
single S3 URL.  As yet there is no database migration support.