Stop using a subapp for better error handling
When the main code used a subapp, exceptions thrown by a handler
would not propagate properly to the main app, instead resulting
in a 500 error with body "Internal Server Error."  This broke
both error reporting in the application logs and error reporting
in failing tests.

Stop using a subapp to avoid this problem, and add a test that
ensures that exceptions from inside external handlers propagate
correctly up to the test level.