Remove unnecessary exception handling on dispatch
The only exceptions we should be getting from dispatch are
serious internal errors like Dramatiq being down or a coding bug,
and letting the raw exception be thrown seems better for those.
We were previously catching and logging them explicitly because
the subapp was eating our exception, but that's been fixed by not
using subapps.