Stop using if TYPE_CHECKING
if TYPE_CHECKING tends to hide errors (there was one bug in the
typing-only includes uncovered by this change) and is just one
more thing one has to think about when writing code.  Stop using
it and import the necessary types directly.  There may be a minor
slowdown during startup, but that's fine, it won't matter.