Add more logging
Add an auth_logger_dependency that binds the user value to the
logger so that we don't have to do it manually.  Add logging to
every state-changing route, and add some logging to the UWS job
processing hooks and the cutout worker so that we get some useful
information in their logs.