Change the exposed URL to /api/cutout
In the current Science Platform URL scheme, the cutout service is
exposed as /api/cutout.  This previously was done with a rewrite
at the ingress layer to avoid encoding the /api path in the
application, since we anticipate moving the API Aspect to its own
domain name eventually.  However, this means that internal URLs
embedded in result documents are incorrect, particularly affecting
the /api/cutout/capabilities endpoint.

To simplify things, add /api as a prefix to the configured routes.
We can always remove it later if needed.