Move more app initialization to top level
Don't initialize the middleware on every startup event, or we
stack multiple copies of the middleware over the course of running
the test suite.  Instead, do this once at the top level since the
middleware doesn't require any configuration.

Do the same for the UWS error handlers (although this probably
doesn't have the same stacking problem).