Enable clipping of the cutout stencil
We originally wanted to raise an error if the cutout stencil didn't
fall within the bounds of the image, but some practical experience
with the Portal has shown that it's better if the stencil clips at
the edge of the image.  This also makes it easier to, for example,
take circular cutouts around a specific object in lots of separate
images that may have different geometries.