Upgrade safir
Upgrade safir to get auth_delegated_token_dependency.

FastAPI had to be upgraded as a result.  It appears that it no longer automatically coerces single-item form parameters to list[str], so some types had to be updated.  (These appear to be intended only as documentation, but are still validated by FastAPI.)