Update dependencies
Add the tox-uv plugin so that the tox virtual environments will also
be managed with uv, and update all Python dependencies.

Update the Debian base image for the Docker container, and update
the cutout worker container to the latest weekly, picking up a new
Butler client.

Fix various tests to remove a deprecation in httpx.