Update dependencies and tox-docker
Update dependencies and separate out the tox dependencies so that
they can be pinned. Update to the current version of tox-docker.

tox-docker no longer uses predictable hostnames or ports for Docker
services, so adjust the configuration parsing and test suite to pick
up the environment variables it injects with the host and port
information.

Convert the PostgreSQL URL to a type based on a string to avoid
having to explicitly convert it to a string. Change the Redis
configuration option to specify a URL instead of a bare hostname and
validate that URL using Pydantic types.

Handle None for the current message, which should be impossible when
running inside Dramatiq (but mypy doesn't know that).