Convert to arq
Switch the task queuing system from Dramatiq to arq. Restructure
the way workers are set up and executed to match the way that arq
thinks of the world instead of the way that Dramatiq thinks of the
world.

This allows task queuing to be async, allowing it to not block the
event loop.

Add new functions, build_worker and build_uws_worker, that construct
the arq worker configuration for the main worker and the UWS database
helper workers.