Add helper class for UWS tests
Several tests have to manually push queued jobs through their
lifecycle in the mock arq queue. Add a helper class that wraps the
data structures required to do this and provides a simple interface.
This can eventually move into Safir when we move the rest of the
support code there.