Drop pydantic-settings from worker, add timeout
Add a separate, minimal UWSWorkerConfig dataclass containing the
information needed to construct the backend worker, and modify the
worker initialization so that it no longer needs to load the main
vo-cutouts Config and thus doesn't depend on pydantic-settings.

Plumb execution_duration in the UWS configuration through to the
arq worker configuration so that we at least impose the maximum
execution duration limit on worker execution. Supporting the per-job
configuration will require more work.