Improve serialization of task errors
Derive `TaskError` from `SlackException` so that eventually we can
report unexpected errors to Slack. Rely on `SlackException` to handle
serialization support, which is currently on a Safir branch.

Add support for including a traceback of the cause of a `TaskError`
exception, and enable that for exceptions from the image cutout
backend. Hopefully those exceptions will not contain any
security-sensitive information.