Refactor the UWS support library
Rework the way that code is divided between the service and the
UWS library to try to move more of the complexity into the library.

The service no longer has to define any routes; instead, it defines
FastAPI dependencies that return the job parameters, and all the
route code is handled by the UWS library.

The policy layer has been removed in favor of adding callbacks to
the UWS configuration layer that accomplish the same thing.  This
refactor standardizes the arguments to the backend worker and passes
it a Pydantic model, since we can rely on Pydantic being available
in stack containers.

Refactor the parameter models for cutouts to support Pydantic
serialization so that there's no need to implement a separate
to_dict function.