Reflect the lack of job cancellation support
Unfortunately, there is no way to cancel a sync worker job within
arq. ThreadPoolWorker does not offer this functionality, nor does
ProcessPoolWorker. We would have to implement our own execution
protocol in an external process that we could kill, which would
pose cleanup problems.

Reflect this by removing the configuration option for setting a job
timeout and forcing the UWS executionduration to 0 by default unless
the service supplies a validator for it (in case there are services
where the sync backend supports an explicit timeout parameter).

Also remove the vestiges of support for aborting a job, since there
is no way to do that either given our worker implementation.