Stop using db_session_dependency in UWS
db_session_dependency provided by Safir is by definition a singleton.
Previously, the UWS layer used it to talk to the UWS database, but
a service using the UWS layer may want its own database connection
for other purposes and the UWS layer was monopolizing it. Fix this
by creating a separate database engine and session inside the UWS
layer.