Separate API model from worker domain model
Rather than using a single set of models for both the API request
and the backend workers, separate them and put most of the heavy
lifting into the API model. Keep the domain model as simple as
possible to avoid pushing dependencies into the worker code.

As long as we're only using the old IVOA protocol, this doesn't
matter that much, but any new JSON-based protocol will want a proper
model for the input parameters, not the version that uses astropy
types and therefore would not generate good API documentation.