Finish the worker and API model separation
The previous change defined separate worker and API models, but still
passed the API model to the backend worker even though it now expected
the worker model. Finish the conversion, which requires a bit more
typing work and a new required method for ParametersModel. Add tests
that the arguments to the backend worker are what we expect them to be.