Add timeout support
Switch the backend worker to ProcessPoolExecutor and use a really
ugly hack to terminate the worker process and rebuild the pool on
timeout or job cancellation.

Restore support for execution duration, change the default execution
duration back to 10 minutes, and re-add the CUTOUT_TIMEOUT
environment variable setting to change the default timeout.