Switch to vo-models where possible for UWS
vo-models provides pydantic-xml models for, among other things, UWS.
This supports serialization to XML without requiring a separate
templating implementation. Use these models for UWS handler output
where possible and remove the corresponding templates.

This as yet doesn't support the handlers that return the full job and
the parameters, since the cutout parameters are multivalued and this
currently apparently cannot be modeled by vo-models.

The corresponding tests now test equality of the pydantic-xml models
instead of the XML text, but continue to use the XML from previous
tests to compare against.