Move PostgreSQL URL manipulation into type
Create a new EnvAsyncPostgresDsn type that modifies the URL based
on environment variable settings set by tox-docker, avoiding the
need for an explicit validator and making it easier to lift this
code into Safir.

Due to a bug in the type system of Pydantic, this does not work in
conjunction with the pattern to convert URLs into strings, so undo
that change and keep the URL as a Pydantic MultiHostUrl object.
This requires explicitly converting it to a string in a few places.