Convert to Wobby and the new Safir UWS library
Remove the direct database connection and convert to the new Safir
UWS library and the Wobbly backend for job storage. Remove the
command-line tool, since its only purpose was to manage the database
schema. Simplify the application startup now that schema creation is
not required.

Change the dependencies to return the full cutout model, now that
Safir supports this. Move the parsing from strings into a helper
function used by the dependencies, since the cutout model no longer
has to be aware of the SODA input parameters.

Move the `MultiValuedParameterError` exception here from Safir, since
that error code is SODA-specific.