Plumb new cutout mode parameter through
Add a new `CUTOUTMODE` parameter to the SODA service that sets the
cutout mode to the three values supported by the backend. Set the
default to `image` to match the new, much faster default. Rename the
`masked_image` setting to `masked-image` to follow our standard style
of using dashes rather than underscores.